feat: converted scripted dashboard. broken browser dashboard
