+++ src/queries/avgRequestSuccessRate.ts
@@ -1,0 +1,18 @@
+import { DSQuery } from 'queries/queries.types';
+
+export function getAvgRequestSuccessRateQuery(label: string): DSQuery {
+  const expr = `
+    avg_over_time(
+      (
+        sum by (${label}, method) (probe_http_requests_failed_total{job="$job", instance="$instance", probe=~"$probe"})
+        /
+        sum by (${label}, method) (probe_http_requests_total{job="$job", instance="$instance", probe=~"$probe"})
+      )[$__range:]
+    )
+  `;
+
+  return {
+    expr,
+    queryType: 'instant',
+  };
+}
