+++ src/scenes/Common/AssertionsTable.types.ts
@@ -1,0 +1,6 @@
+export interface DataRow {
+  name: string;
+  successRate: number;
+  successCount: number;
+  failureCount: number;
+}
