--- src/scenes/SCRIPTED/index.ts
@@ -1,1 +1,0 @@
-export { getScriptedScene } from './scriptedScene';
