+++ src/scenes/Scripted/ResultByTargetTable.utils.ts
